requirements.txt
django==2.2.10
envparse==0.2.0
gunicorn==20.0.4
psycopg2-binary==2.8.6
django-cors-headers==3.7.0
djangorestframework==3.12.3
djangorestframework-simplejwt==4.6.0